refactor: create AbstractAggregationRequestIT class + fail with OpenSearchException details if thrown (#407)
